requirements.txt
# Core dependencies for SAR UAV Human Detection System

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# YOLO (Ultralytics)
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Data handling
PyYAML>=6.0
pandas>=2.0.0

# MAVLink / UAV Integration
pymavlink>=2.4.37
# Note: pymavlink may require additional dependencies on some systems
# For Windows: may need to install via pip install pymavlink
# For Linux: may need python3-dev and other build tools

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: For advanced tracking (if needed)
# filterpy>=1.4.5  # For Kalman filters (already implemented in tracker.py)
# scipy>=1.10.0     # For advanced filtering

# Development (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0

